chore: update ipfs (#43)
* chore: update ipfs

The next ipfs release will be a dual ESM/CJS version which is presenting
some interesting challenges for all the different tools used here.

We also use the `ipfs` module everywhere when we could be using `ipfs-core`
instead - it's lighter and much quicker to install.  Really we should only
use `ipfs` when we need the CLI and the HTTP RPC API, though I'm a little
wary of the potential confusion for new users having two options introduces.